requirements.txt
Flask==3.1.0
google-generativeai==0.8.4
google-auth==2.38.0
requests==2.32.3
